requirements.txt
disnake
requests
requests-cache
python-dotenv
bs4
markdownify
